rmq: added fakes_test.go
