+++ rmq/connection_test.go
@@ -10,40 +10,6 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-type fakeAmqpConn struct {
-	ch                   *amqp.Channel
-	calls                int
-	noOfCallsToReturnErr int
-	err                  error
-}
-
-func (f *fakeAmqpConn) Channel() (*amqp.Channel, error) {
-	f.calls++
-	if f.calls > f.noOfCallsToReturnErr {
-		return f.ch, nil
-	}
-
-	return f.ch, f.err
-}
-
-func (f *fakeAmqpConn) Close() error {
-	// dont really care about this.
-	return nil
-}
-
-func (f *fakeAmqpConn) NotifyClose(receiver chan *amqp.Error) chan *amqp.Error {
-	// similar we dont use this return value for now.
-	return nil
-}
-
-type fakeDialler struct {
-	amqpConn AmqpConn
-}
-
-func (f *fakeDialler) DialConfig(url string, config amqp.Config) (AmqpConn, error) {
-	return f.amqpConn, nil
-}
-
 func Test_NewChannel(t *testing.T) {
 	var testCases = []struct {
 		desc                     string
